Add legacy include guards to public header files
For every public header file, the old include guard definition needs
to be added in addition to the new one

include/openssl/file.h:

    #ifndef __OPENSSL_FILE_H__
    # define __OPENSSL_FILE_H__
    # pragma once

    # define HEADER_FILE_H /* deprecated in version 3.0 */
    ...

This is going to ensure that applications which use the old include guards
externally, for example like this

    #ifndef HEADER_FILE_H
    # include <openssl/file.h>
    #endif

will not fail to compile.

Actually, it would be nice to have a OPENSSL_API_3 deprecation guard
around the legacy include guards, but that would require the inclusion
of <openssl/opensslconf.h> and make automated insertion more tricky.

In addition to the legacy guard, the public header files also receive a
'# pragma once' directive.