Fix header file include guard names
Make the include guards consistent by renaming them systematically according
to the naming conventions below

For the files in the 'include' directory, the guard names match very closely
the path specified in the include directives, with all letters converted to
upper case and '/' and '.' replaced by '_'. For all except the public header
files, an extra 'OSSL' is added as prefix.

   Location  |     Include Directive      |       Include Guard
=============+============================+=================================
include      |                            |
  + openssl  | #include <openssl/file.h>  | #ifndef __OPENSSL_FILE_H__
  |          |                            |
  + internal | #include "internal/file.h" | #ifndef __OSSL_INTERNAL_FILE_H__
  |          |                            |
  + crypto   | #include "crypto/file.h"   | #ifndef __OSSL_CRYPTO_FILE_H__
  |          |                            |
  + ..etc..  |                            |

For the header files in the source tree the naming conventions are similar, but
not followed so rigidly. Sometimes parts of the relative path are used as
constituents of the include guard name for clarity.