squash! test/drbgtest: improve the reseed after fork test
(add the following text to the commit message)

The analysis of #12377 (see [1]) showed that due to an error in the resetting
of the AES-CTR (issue #12405, fixed by #12413), it could happen that only the
first n bytes (n=1,...15) of the children's random output were identical.

This test is optimized to detect this issue by only comparing the first byte
of the sampled data (i.e., the first 'column' of the output). The number of
samples is chosen high enough to keep the chance of false positives low.
The test is executed sixteen times, each time advancing the internal counter
by requesting a single extra byte of random data.

[1] https://github.com/openssl/openssl/issues/12377#issuecomment-656207334